fix: thread safety in cover callbacks
there was a thread safety issue where if a query timed out, the callback
would access values that had gone out of scope in the queueing thread.
this has been fixed by 1) heap-allocating the query user data 2) using
`std::shared_ptr` for the mutex and condition variable. it is now
necessary to keep track of where the user data should be freed.
if there is a timeout, this falls on the callback, for then the queueing
thread has moved on.