config: be smarter about loading config
previously we tried to assign the configuration in its entirety from the
JSON stored by DeadBeeF. it is smarter to construct the default
configuration and merge the DeaDBeeF JSON into it. this means that the
introduction of a new configuration variable, such as the timeout
introduced in the previous commit, do not reset the user's configuration
to the default. instead, other values will retain their values and the
new variable will have its default value.