gui: fix memory management in `ProgressMonitor`
previously we had a `std::shared_ptr<std::atomic<NotifyNotification*>>`
passing `g_object_unref` as a deleter. this was incorrect for several
reasons. first, calling `store(nullptr)` on the managed atomic does not
trigger the deleter, because the object isn't replaced. second,
`g_object_unref` will be called with the address of an atomic, not the
stored value. third, `std::atomic` isn't strong enough synchronisation;
we need the `NotifyNotification*` to remain valid after reading its
value. additionally, the several layers of indirection were hard to
understand.

the new model is that the `ProgressMonitor` instance owns the
`NotifyNotification` and unrefs it in the destructor. access to it is
protected by a mutex. the signal handler for the notification being
closed takes a pointer to the owning instance as user data, which is
safe because the instance always outlives the notification.