feat: reference count synced files in the database
this requires some modifications of `queue_jobs` and `queue_cover_jobs` so
that they can handle the case of the same file appearing in multiple
playlists, which needs to be recorded, but shouldn't cause multiple jobs
to queued. in particular, `queue_cover_jobs` now takes a `path ->
struct` map where the struct contains the ddb playlist item and a vector
of playlist uuids where it appears.